Refactor srlinux.py (#2)
* update packages

* Refactor NetworkInstance methods into NetworkInstanceMixin for better code organization

* Add BGP functionality as a mixin and refactor related methods

- Introduced BgpMixin class with methods for retrieving BGP RIB and summary information.
- Extracted BGP-related methods from srlinux.py to bgp.py for better organization and separation of concerns.
- Updated SrLinux class to inherit from BgpMixin, enabling BGP functionality.
- Refactored get_bgp_rib and get_sum_bgp methods to utilize the new BgpMixin.
- Enhanced error handling and data processing within BGP methods.
- Improved code readability and maintainability by restructuring complex logic into helper functions.

* satisfy mypy

* add an example

* renamed files

* Add mixins for Layer2, Neighbor Discovery, and System functionalities; refactor routing mixin for improved organization
